squash: move the tap logic to util
* This moves the tap logic out of cmd.js and into a separate function in utils
