+++ lib/utils.js
@@ -1,9 +1,12 @@
 'use strict'
 
+const fs = require('fs')
 const exec = require('child_process').exec
 const http = require('http')
 const https = require('https')
 const url = require('url')
+const formatTap = require('../lib/format-tap')
+const Tap = require('../lib/tap')
 const pretty = require('../lib/format-pretty')
 const chalk = require('chalk')
 const CHECK = chalk.green('✔')
@@ -119,6 +122,39 @@ exports.validateMetadata = function(validator, args) {
     exports.load(sha, (err, data) => {
       if (err) throw err
       validator.lint(data)
+    })
+  }
+
+  run()
+}
+
+exports.parseTap = function(validator, parsed, args) {
+  const tap = new Tap()
+  tap.pipe(process.stdout)
+  if (parsed.out) tap.pipe(fs.createWriteStream(parsed.out))
+  let count = 0
+  let total = args.length
+
+  validator.on('commit', (c) => {
+    count++
+    const test = tap.test(c.commit.sha)
+    formatTap(test, c.commit, c.messages, validator)
+    if (count === total) {
+      setImmediate(() => {
+        tap.end()
+        if (tap.status === 'fail')
+          process.exitCode = 1
+      })
+    }
+  })
+
+  function run() {
+    if (!args.length) return
+    const sha = args.shift()
+    exports.load(sha, (err, data) => {
+      if (err) throw err
+      validator.lint(data)
+      run()
     })
   }
